+++ apps/cowswap-frontend/src/modules/twap/containers/EoaTwapSigningPendingContent/EoaTwapSigningPendingContent.tsx
@@ -1,14 +1,9 @@
-import { ReactNode, useCallback } from 'react'
+import { ReactNode } from 'react'
 
-import { useMediaQuery } from '@cowprotocol/common-hooks'
 import { getExplorerTwapOrderLink } from '@cowprotocol/common-utils'
-import { Media } from '@cowprotocol/ui'
 import { useWalletInfo } from '@cowprotocol/wallet'
 
-import { OrderTabId } from 'entities/routes/routes.atom'
-
-import { useRevealOrderInOrdersTable } from 'modules/ordersTable'
-import { OrderStep, OrderSteps, useSetOrdersTableDrawerOpen } from 'modules/trade'
+import { OrderStep, OrderSteps } from 'modules/trade'
 import { TradeFormBlankButton } from 'modules/tradeFormValidation'
 
 import { EoaTwapSuccessContent } from './EoaTwapSuccessContent.pure'
@@ -20,33 +15,16 @@ import { EoaTwapCurrentStepButtonProps } from '../../utils/buildEoaTwapConfirmat
 export interface EoaTwapSigningPendingContentProps {
   steps: OrderStep[]
   buttonProps: EoaTwapCurrentStepButtonProps | null
-  onDismiss(): void
+  onViewOrders(): void
 }
 
 export function EoaTwapSigningPendingContent({
   steps,
   buttonProps,
-  onDismiss,
+  onViewOrders,
 }: EoaTwapSigningPendingContentProps): ReactNode {
   const signingStep = useEoaTwapSigningStep()
   const { chainId } = useWalletInfo()
-  const revealOrderInOrdersTable = useRevealOrderInOrdersTable()
-  const setOrdersTableDrawerOpen = useSetOrdersTableDrawerOpen()
-  const isUpToLarge = useMediaQuery(Media.upToLarge(false))
-
-  const onViewOrders = useCallback(async () => {
-    const orderId = signingStep?.eventId
-
-    onDismiss()
-
-    if (isUpToLarge) {
-      setOrdersTableDrawerOpen(true)
-    }
-
-    if (orderId) {
-      await revealOrderInOrdersTable(orderId, OrderTabId.OPEN)
-    }
-  }, [isUpToLarge, onDismiss, revealOrderInOrdersTable, setOrdersTableDrawerOpen, signingStep?.eventId])
 
   if (!signingStep) return null
 
@@ -54,7 +32,7 @@ export function EoaTwapSigningPendingContent({
     const explorerUrl =
       chainId && signingStep.eventId ? getExplorerTwapOrderLink(chainId, signingStep.eventId) : undefined
 
-    return <EoaTwapSuccessContent explorerUrl={explorerUrl} onNewTrade={onDismiss} onViewOrders={onViewOrders} />
+    return <EoaTwapSuccessContent explorerUrl={explorerUrl} onViewOrders={onViewOrders} />
   }
 
   return (